refactor(reposicao): extrai leaf components de AdminReposicaoPromocaoDetail (1797→1205)
Primeiro passo do split do god-component #1 (era 1797 LoC). Extrai as
peças auto-contidas pra src/components/reposicao/promocaoDetail/, sem
mudança de lógica (move + import):

- types.ts        — Campanha, ItemRow, ItemEfetivo, Evento + consts
                    (EMPRESA, FORNECEDOR_DEFAULT, ESTADO_LABEL,
                    TIPO_EVENTO_LABELS)
- helpers.tsx     — tipoEventoIcon, estadoBadgeClass, confiancaBadge,
                    formatDateTimeBR, formatRelative
- MapeamentoStatusCell.tsx — célula de status do mapeamento SKU (~348L)
- DescontoExtraCell.tsx    — popover de desconto extra (~137L)

Página: 1797 → 1205 LoC (-592). Remove 9 imports que ficaram órfãos
(Send, Reply, AlertCircle, StickyNote, Search, Checkbox, Popover*).

Os 4 arquivos novos já entram no tsconfig.strict.json (totalmente
tipados). Strict: 206 → 210.

Verificado: typecheck:strict exit 0, bunx tsc --noEmit limpo, 344
testes passando, bun run build OK.
