+++ src/pages/AdminReposicaoPromocaoDetail.tsx
@@ -15,12 +15,7 @@ import {
   Trash2,
   Check,
   X,
-  Send,
-  Reply,
-  AlertCircle,
-  StickyNote,
   Sparkles,
-  Search,
 } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
@@ -28,7 +23,6 @@ import { Badge } from "@/components/ui/badge";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import { Label } from "@/components/ui/label";
-import { Checkbox } from "@/components/ui/checkbox";
 import {
   Tabs,
   TabsContent,
@@ -50,11 +44,6 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
 import {
   Tooltip,
   TooltipContent,
@@ -79,606 +68,25 @@ import {
   AlertDialogHeader,
   AlertDialogTitle,
 } from "@/components/ui/alert-dialog";
-
-const EMPRESA = "OBEN";
-const FORNECEDOR_DEFAULT = "RENNER SAYERLACK S/A";
-
-type Campanha = {
-  id: number;
-  empresa: string;
-  nome: string;
-  fornecedor_nome: string;
-  tipo_origem: string;
-  data_inicio: string;
-  data_fim: string;
-  estado: string;
-  observacoes: string | null;
-  origem_arquivo_url: string | null;
-  origem_arquivo_tipo: string | null;
-  origem_email_assunto: string | null;
-  origem_email_remetente: string | null;
-  origem_email_data: string | null;
-  extracao_confianca: number | null;
-  extracao_observacoes: string | null;
-  extraido_em: string | null;
-  criado_em: string;
-};
-
-type ItemRow = {
-  id: number;
-  campanha_id: number;
-  sku_codigo_fornecedor: string;
-  descricao_produto_fornecedor: string | null;
-  sku_codigo_omie: number | null;
-  mapeamento_qualidade: string | null;
-  mapeamento_candidatos: unknown;
-  desconto_perc: number;
-  volume_minimo: number | null;
-  confirmado: boolean;
-  ativo: boolean;
-  desconto_extra_perc: number | null;
-  desconto_extra_observacoes: string | null;
-  desconto_extra_negociado_por: string | null;
-  desconto_extra_negociado_em: string | null;
-  desconto_extra_email_referencia: string | null;
-};
-
-type ItemEfetivo = {
-  id: number;
-  desconto_efetivo: number;
-};
-
-type Evento = {
-  id: number;
-  campanha_id: number;
-  tipo_evento: string;
-  desconto_perc_proposto: number | null;
-  volume_minimo_proposto: number | null;
-  data_evento: string;
-  email_referencia: string | null;
-  conteudo: string | null;
-  registrado_por: string | null;
-  registrado_em: string;
-};
-
-const TIPO_EVENTO_LABELS: Record<string, string> = {
-  proposta_enviada: "Proposta enviada",
-  contraproposta_recebida: "Contra-proposta recebida",
-  aceite_lucas: "Aceite (Lucas)",
-  aceite_gerente: "Aceite (Gerente)",
-  recusa_gerente: "Recusa (Gerente)",
-  abandono: "Abandono",
-  nota: "Nota",
-};
-
-function tipoEventoIcon(tipo: string) {
-  switch (tipo) {
-    case "proposta_enviada":
-      return Send;
-    case "contraproposta_recebida":
-      return Reply;
-    case "aceite_lucas":
-    case "aceite_gerente":
-      return Check;
-    case "recusa_gerente":
-      return X;
-    case "abandono":
-      return AlertCircle;
-    case "nota":
-    default:
-      return StickyNote;
-  }
-}
-
-function estadoBadgeClass(estado: string): string {
-  switch (estado) {
-    case "rascunho":
-      return "bg-status-warning/15 text-status-warning border-status-warning/30";
-    case "negociando":
-      return "bg-status-info/15 text-status-info border-status-info/30";
-    case "ativa":
-      return "bg-status-success/15 text-status-success border-status-success/30";
-    case "encerrada":
-      return "bg-muted text-muted-foreground border-border";
-    case "cancelada":
-      return "bg-destructive/15 text-destructive border-destructive/30";
-    default:
-      return "";
-  }
-}
-
-const ESTADO_LABEL: Record<string, string> = {
-  rascunho: "Rascunho",
-  negociando: "Negociando",
-  ativa: "Ativa",
-  encerrada: "Encerrada",
-  cancelada: "Cancelada",
-};
-
-function confiancaBadge(c: number | null) {
-  if (c === null || c === undefined) return null;
-  let cls =
-    "bg-status-success/15 text-status-success border-status-success/30";
-  if (c < 0.5) cls = "bg-destructive/15 text-destructive border-destructive/30";
-  else if (c <= 0.8)
-    cls =
-      "bg-status-warning/15 text-status-warning border-status-warning/30";
-  return (
-    <Badge variant="outline" className={cls}>
-      Confiança {Math.round(c * 100)}%
-    </Badge>
-  );
-}
-
-function formatDateTimeBR(iso: string): string {
-  const d = new Date(iso);
-  return d.toLocaleString("pt-BR", {
-    day: "2-digit",
-    month: "2-digit",
-    year: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-  });
-}
-
-function formatRelative(iso: string): string {
-  const d = new Date(iso);
-  const diffMs = Date.now() - d.getTime();
-  const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-  if (diffDays === 0) return "hoje";
-  if (diffDays === 1) return "há 1 dia";
-  if (diffDays < 30) return `há ${diffDays} dias`;
-  const diffMonths = Math.floor(diffDays / 30);
-  if (diffMonths === 1) return "há 1 mês";
-  return `há ${diffMonths} meses`;
-}
-
-// ========== STATUS BADGE DO MAPEAMENTO ==========
-function MapeamentoStatusCell({
-  item,
-  onUpdate,
-}: {
-  item: ItemRow;
-  onUpdate: (changes: Partial<ItemRow>) => void;
-}) {
-  const qc = useQueryClient();
-  const [searchOpen, setSearchOpen] = useState(false);
-  const [searchQuery, setSearchQuery] = useState("");
-  const [searchResults, setSearchResults] = useState<
-    Array<{ omie_codigo_produto: number; descricao: string; codigo: string }>
-  >([]);
-  const [searching, setSearching] = useState(false);
-  const [selectedSkus, setSelectedSkus] = useState<Record<number, { descricao: string; codigo: string }>>({});
-  const [salvando, setSalvando] = useState(false);
-
-  // Busca debounced
-  useEffect(() => {
-    if (!searchOpen || searchQuery.trim().length < 2) {
-      setSearchResults([]);
-      return;
-    }
-    const t = setTimeout(async () => {
-      setSearching(true);
-      const term = searchQuery.trim();
-      // Busca por descrição OU código OU sku omie (numérico). account no banco é lowercase.
-      const isNumeric = /^\d+$/.test(term);
-      type OmieSearchRow = { omie_codigo_produto: number; descricao: string; codigo: string };
-      let query = supabase
-        .from("omie_products")
-        .select("omie_codigo_produto, descricao, codigo")
-        .eq("account", EMPRESA.toLowerCase())
-        .eq("ativo", true);
-      query = isNumeric
-        ? query.or(
-            `codigo.ilike.%${term}%,omie_codigo_produto.eq.${term}`,
-          )
-        : query.or(
-            `descricao.ilike.%${term}%,codigo.ilike.%${term}%`,
-          );
-      const { data, error } = await query.limit(20);
-      setSearching(false);
-      if (!error) setSearchResults((data as unknown as OmieSearchRow[]) || []);
-    }, 300);
-    return () => clearTimeout(t);
-  }, [searchQuery, searchOpen]);
-
-  const q = item.mapeamento_qualidade;
-  const isConfirmed = item.confirmado;
-
-  // ========== unico + confirmado ==========
-  if (q === "unico" && isConfirmed) {
-    return (
-      <TooltipProvider>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <Badge
-              variant="outline"
-              className="bg-status-success/15 text-status-success border-status-success/30 cursor-default"
-            >
-              <Check className="h-3 w-3 mr-1" /> OK
-            </Badge>
-          </TooltipTrigger>
-          <TooltipContent className="max-w-xs">
-            <div className="space-y-1 text-xs">
-              <div className="font-mono">{item.sku_codigo_omie}</div>
-              <div>{item.descricao_produto_fornecedor || "—"}</div>
-            </div>
-          </TooltipContent>
-        </Tooltip>
-      </TooltipProvider>
-    );
-  }
-
-  // ========== expandido_automatico + confirmado ==========
-  if (q === "expandido_automatico" && isConfirmed) {
-    return (
-      <TooltipProvider>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <Badge
-              variant="outline"
-              className="bg-status-success/15 text-status-success border-status-success/30 cursor-default"
-            >
-              <Sparkles className="h-3 w-3 mr-1" /> Variante
-            </Badge>
-          </TooltipTrigger>
-          <TooltipContent className="max-w-xs">
-            <div className="space-y-1 text-xs">
-              <div className="font-mono">{item.sku_codigo_omie}</div>
-              <div>{item.descricao_produto_fornecedor || "—"}</div>
-              <div className="text-muted-foreground italic">
-                Expandido automaticamente de {item.sku_codigo_fornecedor}
-              </div>
-            </div>
-          </TooltipContent>
-        </Tooltip>
-      </TooltipProvider>
-    );
-  }
-
-  // ========== similaridade (precisa revisão) ==========
-  if (
-    (q === "unico_por_similaridade" || q === "expandido_por_similaridade") &&
-    !isConfirmed
-  ) {
-    return (
-      <Popover>
-        <PopoverTrigger asChild>
-          <Badge
-            variant="outline"
-            className="bg-status-warning/15 text-status-warning border-status-warning/30 cursor-pointer"
-          >
-            Revisar — similaridade
-          </Badge>
-        </PopoverTrigger>
-        <PopoverContent className="w-80">
-          <div className="space-y-3">
-            <div>
-              <div className="text-xs text-muted-foreground">SKU Omie</div>
-              <div className="font-mono text-sm">{item.sku_codigo_omie}</div>
-            </div>
-            <div>
-              <div className="text-xs text-muted-foreground">Descrição</div>
-              <div className="text-sm">
-                {item.descricao_produto_fornecedor || "—"}
-              </div>
-            </div>
-            <p className="text-xs italic text-muted-foreground">
-              Resolvido por busca aproximada. Confirme se está correto.
-            </p>
-            <Button
-              size="sm"
-              className="w-full"
-              onClick={() => onUpdate({ confirmado: true })}
-            >
-              <Check className="h-4 w-4" /> Confirmar
-            </Button>
-          </div>
-        </PopoverContent>
-      </Popover>
-    );
-  }
-
-  // ========== manual_confirmado ==========
-  if (q === "manual_confirmado" && isConfirmed) {
-    return (
-      <TooltipProvider>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <Badge
-              variant="outline"
-              className="bg-status-info/15 text-status-info border-status-info/30 cursor-default"
-            >
-              Manual
-            </Badge>
-          </TooltipTrigger>
-          <TooltipContent className="max-w-xs">
-            <div className="space-y-1 text-xs">
-              <div className="font-mono">{item.sku_codigo_omie}</div>
-              <div>{item.descricao_produto_fornecedor || "—"}</div>
-            </div>
-          </TooltipContent>
-        </Tooltip>
-      </TooltipProvider>
-    );
-  }
-
-  // ========== nao_encontrado (ou null) — busca manual com sugestão automática ==========
-  // Pré-popula o campo com a descrição do fornecedor — 99% dos casos bate via ilike.
-  const sugestaoTermo = (() => {
-    const base = (item.descricao_produto_fornecedor || item.sku_codigo_fornecedor || "").trim();
-    if (!base) return "";
-    const palavras = base
-      .replace(/[^\w\sÀ-ÿ.]/g, " ")
-      .split(/\s+/)
-      .filter((w) => w.length >= 3)
-      .slice(0, 3);
-    return palavras.join(" ");
-  })();
-
-  return (
-    <Popover
-      open={searchOpen}
-      onOpenChange={(open) => {
-        setSearchOpen(open);
-        if (open && !searchQuery) setSearchQuery(sugestaoTermo);
-      }}
-    >
-      <PopoverTrigger asChild>
-        <Badge
-          variant="outline"
-          className="bg-destructive/15 text-destructive border-destructive/30 cursor-pointer"
-        >
-          {q === "nao_encontrado" ? "Não encontrado" : "Pendente"}
-        </Badge>
-      </PopoverTrigger>
-      <PopoverContent className="w-[420px]">
-        <div className="space-y-3">
-          <div>
-            <Label className="text-xs">Buscar produto Omie</Label>
-            <p className="text-[11px] text-muted-foreground mb-1">
-              Selecione <strong>uma ou mais embalagens</strong> — útil quando o código fornecedor
-              (ex.: {item.sku_codigo_fornecedor}) cobre vários tamanhos (0,9L · 3,6L · 18L…).
-            </p>
-            <div className="relative mt-1">
-              <Search className="h-4 w-4 absolute left-3 top-1/2 -translate-y-1/2 text-muted-foreground" />
-              <Input
-                placeholder="Buscar por descrição, código ou SKU Omie…"
-                className="pl-9"
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-                autoFocus
-              />
-            </div>
-          </div>
-          <div className="max-h-64 overflow-y-auto space-y-1">
-            {searching && (
-              <div className="text-xs text-muted-foreground flex items-center gap-2 p-2">
-                <Loader2 className="h-3 w-3 animate-spin" /> Buscando…
-              </div>
-            )}
-            {!searching && searchQuery.length >= 2 && searchResults.length === 0 && (
-              <div className="text-xs text-muted-foreground p-2">
-                Nenhum produto encontrado
-              </div>
-            )}
-            {searchResults.map((p) => {
-              const checked = !!selectedSkus[p.omie_codigo_produto];
-              return (
-                <label
-                  key={p.omie_codigo_produto}
-                  className="flex items-start gap-2 w-full text-left p-2 rounded hover:bg-accent transition-colors cursor-pointer"
-                >
-                  <Checkbox
-                    checked={checked}
-                    onCheckedChange={(v) => {
-                      setSelectedSkus((prev) => {
-                        const next = { ...prev };
-                        if (v) next[p.omie_codigo_produto] = { descricao: p.descricao, codigo: p.codigo };
-                        else delete next[p.omie_codigo_produto];
-                        return next;
-                      });
-                    }}
-                    className="mt-0.5"
-                  />
-                  <div className="flex-1 min-w-0">
-                    <div className="text-xs font-mono text-muted-foreground">
-                      {p.codigo} · {p.omie_codigo_produto}
-                    </div>
-                    <div className="text-sm">{p.descricao}</div>
-                  </div>
-                </label>
-              );
-            })}
-          </div>
-          <div className="flex items-center justify-between border-t pt-2">
-            <span className="text-xs text-muted-foreground">
-              {Object.keys(selectedSkus).length} selecionado(s)
-            </span>
-            <Button
-              size="sm"
-              disabled={Object.keys(selectedSkus).length === 0 || salvando}
-              onClick={async () => {
-                const ids = Object.keys(selectedSkus).map(Number);
-                if (ids.length === 0) return;
-                setSalvando(true);
-                try {
-                  // Primeiro SKU → atualiza o item original (in-place)
-                  const primeiroId = ids[0];
-                  const primeiro = selectedSkus[primeiroId];
-                  onUpdate({
-                    sku_codigo_omie: primeiroId,
-                    descricao_produto_fornecedor: primeiro.descricao,
-                    mapeamento_qualidade: "manual_confirmado",
-                    confirmado: true,
-                  });
-
-                  // Demais → inserir novos itens irmãos (mesmo desconto/volume).
-                  // Para escapar do unique (campanha_id, sku_codigo_fornecedor, volume_minimo),
-                  // sufixamos o código fornecedor com #omie<id>.
-                  const extras = ids.slice(1);
-                  if (extras.length > 0) {
-                    const payload = extras.map((omieId) => ({
-                      campanha_id: item.campanha_id,
-                      sku_codigo_fornecedor: `${item.sku_codigo_fornecedor}#omie${omieId}`,
-                      descricao_produto_fornecedor: selectedSkus[omieId].descricao,
-                      sku_codigo_omie: omieId,
-                      mapeamento_qualidade: "manual_confirmado",
-                      desconto_perc: item.desconto_perc,
-                      volume_minimo: item.volume_minimo,
-                      confirmado: true,
-                      ativo: true,
-                      observacoes: `Expandido manualmente a partir de ${item.sku_codigo_fornecedor}`,
-                    }));
-                    const { error } = await supabase
-                      .from("promocao_item")
-                      .insert(payload as never);
-                    if (error) throw error;
-                    toast.success(`${ids.length} embalagens vinculadas`);
-                  } else {
-                    toast.success("Produto vinculado");
-                  }
-                  qc.invalidateQueries({ queryKey: ["promocao-itens", String(item.campanha_id)] });
-                  setSelectedSkus({});
-                  setSearchOpen(false);
-                } catch (e) {
-                  toast.error(e instanceof Error ? e.message : "Erro ao vincular");
-                } finally {
-                  setSalvando(false);
-                }
-              }}
-            >
-              {salvando && <Loader2 className="h-3 w-3 animate-spin mr-1" />}
-              Confirmar
-            </Button>
-          </div>
-        </div>
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-// ========== POPOVER DESCONTO EXTRA ==========
-function DescontoExtraCell({
-  item,
-  onSave,
-  userEmail,
-}: {
-  item: ItemRow;
-  onSave: (changes: Partial<ItemRow>) => void;
-  userEmail: string;
-}) {
-  const [open, setOpen] = useState(false);
-  const [perc, setPerc] = useState<string>(
-    item.desconto_extra_perc?.toString() || "",
-  );
-  const [obs, setObs] = useState(item.desconto_extra_observacoes || "");
-  const [emailRef, setEmailRef] = useState(item.desconto_extra_email_referencia || "");
-
-  useEffect(() => {
-    if (open) {
-      setPerc(item.desconto_extra_perc?.toString() || "");
-      setObs(item.desconto_extra_observacoes || "");
-      setEmailRef(item.desconto_extra_email_referencia || "");
-    }
-  }, [open, item]);
-
-  const handleSave = () => {
-    const num = parseFloat(perc);
-    if (isNaN(num) || num <= 0 || num > 50) {
-      toast.error("Desconto extra deve ser entre 0 e 50%");
-      return;
-    }
-    onSave({
-      desconto_extra_perc: num,
-      desconto_extra_observacoes: obs || null,
-      desconto_extra_email_referencia: emailRef || null,
-      desconto_extra_negociado_por: userEmail,
-      desconto_extra_negociado_em: new Date().toISOString(),
-    });
-    setOpen(false);
-  };
-
-  const handleRemove = () => {
-    onSave({
-      desconto_extra_perc: null,
-      desconto_extra_observacoes: null,
-      desconto_extra_email_referencia: null,
-      desconto_extra_negociado_por: null,
-      desconto_extra_negociado_em: null,
-    });
-    setOpen(false);
-  };
-
-  const tem = item.desconto_extra_perc !== null;
-
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger asChild>
-        {tem ? (
-          <Badge
-            variant="outline"
-            className="bg-status-info/15 text-status-info border-status-info/30 cursor-pointer"
-          >
-            base + {item.desconto_extra_perc}%
-          </Badge>
-        ) : (
-          <Button variant="ghost" size="sm" className="h-6 px-2 text-xs">
-            <Plus className="h-3 w-3" /> extra
-          </Button>
-        )}
-      </PopoverTrigger>
-      <PopoverContent className="w-80">
-        <div className="space-y-3">
-          <div className="font-medium text-sm">Desconto extra negociado</div>
-          <div>
-            <Label className="text-xs">Percentual extra (%)</Label>
-            <Input
-              type="number"
-              step="0.1"
-              min="0"
-              max="50"
-              value={perc}
-              onChange={(e) => setPerc(e.target.value)}
-              placeholder="Ex: 5"
-            />
-          </div>
-          <div>
-            <Label className="text-xs">Observações</Label>
-            <Textarea
-              value={obs}
-              onChange={(e) => setObs(e.target.value)}
-              placeholder="Negociado com [nome] em [data]"
-              rows={2}
-            />
-          </div>
-          <div>
-            <Label className="text-xs">Referência de email (opcional)</Label>
-            <Input
-              value={emailRef}
-              onChange={(e) => setEmailRef(e.target.value)}
-              placeholder="Assunto ou link"
-            />
-          </div>
-          <div className="flex gap-2">
-            <Button size="sm" onClick={handleSave} className="flex-1">
-              Salvar
-            </Button>
-            {tem && (
-              <Button
-                size="sm"
-                variant="destructive"
-                onClick={handleRemove}
-              >
-                Remover
-              </Button>
-            )}
-          </div>
-        </div>
-      </PopoverContent>
-    </Popover>
-  );
-}
+import {
+  EMPRESA,
+  FORNECEDOR_DEFAULT,
+  ESTADO_LABEL,
+  TIPO_EVENTO_LABELS,
+  type Campanha,
+  type ItemRow,
+  type ItemEfetivo,
+  type Evento,
+} from "@/components/reposicao/promocaoDetail/types";
+import {
+  tipoEventoIcon,
+  estadoBadgeClass,
+  confiancaBadge,
+  formatDateTimeBR,
+  formatRelative,
+} from "@/components/reposicao/promocaoDetail/helpers";
+import { MapeamentoStatusCell } from "@/components/reposicao/promocaoDetail/MapeamentoStatusCell";
+import { DescontoExtraCell } from "@/components/reposicao/promocaoDetail/DescontoExtraCell";
 
 // ========== PÁGINA PRINCIPAL ==========
 export default function AdminReposicaoPromocaoDetail() {
